feat(controller): connect the layers and operate on the data sent from form

+++ src/main/java/edu/teis/appmonedas/dto/IndexFormDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class IndexFormDto {
-    private float importe;
+    private Double importe;
     private String divisaOrigen;
     private String divisaDestino;
